# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "FinZ LMS API",
        "description": "API documentation for FinZ LMS backend",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/v1/admin/blacklist": {
            "get": {
                "tags": [
                    "Blacklist"
                ],
                "summary": "index Blacklist",
                "operationId": "ea923a7b8bb096effd7c2b680c1629b0",
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Blacklist"
                ],
                "summary": "store Blacklist",
                "operationId": "7b8cbff79b0b4f07cb5a4524bb908044",
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/admin/blacklist/bulk-import": {
            "post": {
                "tags": [
                    "Blacklist"
                ],
                "summary": "bulkImport Blacklist",
                "operationId": "b7c83b89922a30231cbd1b151cedca96",
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/admin/blacklist/{id}/remove": {
            "post": {
                "tags": [
                    "Blacklist"
                ],
                "summary": "remove Blacklist",
                "operationId": "d69383d2fbc2e3aee1f6efa053dd65fc",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/admin/blacklist/{id}/whitelist-override": {
            "post": {
                "tags": [
                    "Blacklist"
                ],
                "summary": "whitelistOverride Blacklist",
                "operationId": "dfd34ccbd08a68d20f3ea6f42baa1040",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/api/v1/admin/collections": {
            "get": {
                "tags": [
                    "Collection"
                ],
                "summary": "index Collection",
                "operationId": "2b5198fa43fda394cee2d25a25d75980",
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/admin/collections/{id}/assign-agent": {
            "post": {
                "tags": [
                    "Collection"
                ],
                "summary": "assignAgent Collection",
                "operationId": "780f2e1cc46640ff385553b39d5a98b5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/admin/collections/bounces": {
            "get": {
                "tags": [
                    "Collection"
                ],
                "summary": "bounceFeed Collection",
                "operationId": "2c00d0eca8a14a3dc6950ece362a1709",
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/admin/collections/bounces/{id}/retry": {
            "post": {
                "tags": [
                    "Collection"
                ],
                "summary": "retryBounce Collection",
                "operationId": "35e2e18dccad0aa607eefe8083e0cee3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/admin/collections/{id}/npa-status": {
            "post": {
                "tags": [
                    "Collection"
                ],
                "summary": "setNpaStatus Collection",
                "operationId": "7f80155a656275cbfa3987ecfda8acb0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}